Improved graphics, edited pause mode, added audio
The paddles now have colors, as well as the ball. In the middle of the screen you can now see a dotted line to split the two halfs of the player, like in the original Pong game. When you pause the game with space, you cannot move the paddles anymore in order to prevent cheating (pausing the game and then set the paddle to its correct position). I also added two audio file. When the ball hits a paddle or  a players scores, the corresponding sound is played.